backend/requirements.txt
# Core Dependencies
fastapi==0.116.1
numpy==2.3.2
pixelmatch==0.3.0
pillow==11.3.0
python-multipart==0.0.20
pydantic-settings==2.10.1
uvicorn==0.35.0

# Development and Testing Dependencies
httpx==0.28.1
pytest==8.4.1
